+++ slack/data_slack_users_group.go
@@ -74,11 +74,3 @@ func hashStringSlice(slice []string) string {
 	}
 	return hex.EncodeToString(h.Sum(nil))
 }
-
-func convertStringSliceToInterface(slice []string) []interface{} {
-	result := make([]interface{}, len(slice))
-	for i, v := range slice {
-		result[i] = v
-	}
-	return result
-}